Refactor: Update import paths for type definitions
