+++ packages/giselle-engine/src/core/vector-store/paths.ts
@@ -1,4 +1,4 @@
-import type { VectorStoreId } from "./object";
+import type { VectorStoreId } from "./types/object";
 
 export function vectorStorePath(vectorStoreId: VectorStoreId) {
 	return `vectorStores/${vectorStoreId}/object.json`;
