+++ packages/vector-store-adapters/tsup.config.ts
@@ -4,7 +4,7 @@ export default defineConfig([
 	{
 		entry: ["src/index.ts"],
 		format: ["cjs", "esm"],
-		external: ['zod'],
+		external: ["zod"],
 		dts: true,
 		sourcemap: true,
 	},
